+++ src/Submodules/SeisHDF.jl
@@ -1,7 +1,9 @@
 module SeisHDF
-using Dates, HDF5, SeisIO
+using Dates, HDF5, SeisIO, SeisIO.Quake
+import LightXML: free, parse_string
 import SeisIO: KW, TimeSpec, check_for_gap!, dtconst, endtime, mk_xml!,
   parsetimewin, read_station_xml!, split_id, sxml_mergehdr!, t_win, trunc_x!
+import SeisIO.Quake:event_xml!
 
 # These are adapted from unix2datetime.(1.0e-9.*[typemin(Int64), typemax(Int64)])
 const unset_s = "1677-09-21T00:12:44"
@@ -24,9 +26,10 @@ include("SeisHDF/write_asdf.jl")
 include("SeisHDF/read_hdf5.jl")
 include("SeisHDF/write_hdf5.jl")
 
-# scanner
+# scanners
 include("SeisHDF/scan_hdf5.jl")
+include("SeisHDF/asdf_qml.jl")
 
-export read_hdf5, read_hdf5!, scan_hdf5, write_hdf5
+export asdf_qml, read_hdf5, read_hdf5!, scan_hdf5, write_hdf5
 
 end